refactor(user): Resolver/Service DTO 전환 + 중복 검증 제거
A-2 검증 전략 P0-3 단계 3 (User Input 전수) — 적용편.

- User Resolver 7종이 신규 DTO class 를 @Args 타입으로 사용.
  GraphQL ValidationPipe 가 정식 검증 수행.
- user-{order,wishlist,recent-view,review}.service: offset/limit 수동
  검증 제거 (DTO 가 처리)
- user-review.service: rating · content 길이 수동 검증 제거 (DTO 가
  처리). validateMedia(이미지 10 / 영상 1) invariant 만 잔존.
- user-profile.service: profileImageUrl 의 길이/공백 수동 검증 제거
  (DTO 가 처리). Name 빈 문자열 reject 는 service 에 방어 코드로 유지.
- types/user-{input,order-input,review-input}.type.ts 삭제 (interface
  →  class 마이그레이션 완료, 더 이상 사용처 없음)
- user-wishlist-output.type.ts 에 잘못 위치해 있던 MyWishlistInput 제거
- spec 정리: DTO 레이어로 이전된 검증 케이스(15+) 제거. service spec
  은 도메인 분기에 집중.

FE 영향: 정상 입력은 응답 동일. 잘못된 형식은 ValidationPipe 가
GraphQL BAD_REQUEST 로 일관 응답 (메시지 본문은 기존 한국어 그대로
유지하도록 룰별 message 옵션 명시).
