--- src/features/user/types/user-input.type.ts
@@ -1,28 +1,0 @@
-export interface CompleteOnboardingInput {
-  name?: string | null;
-  nickname: string;
-  birthDate?: Date | null;
-  phoneNumber?: string | null;
-}
-
-export interface UpdateMyProfileInput {
-  nickname?: string | null;
-  name?: string | null;
-  birthDate?: Date | null;
-  phoneNumber?: string | null;
-}
-
-export interface UpdateMyProfileImageInput {
-  profileImageUrl: string;
-}
-
-export interface MyNotificationsInput {
-  unreadOnly?: boolean | null;
-  offset?: number | null;
-  limit?: number | null;
-}
-
-export interface MySearchHistoriesInput {
-  offset?: number | null;
-  limit?: number | null;
-}
