--- src/features/user/types/user-order-input.type.ts
@@ -1,7 +1,0 @@
-import type { OrderStatus } from '@prisma/client';
-
-export interface MyOrdersInput {
-  statuses?: OrderStatus[];
-  offset?: number;
-  limit?: number;
-}
